Pulled the file path finding out of the input line
