+++ src/ScriptCs.Contracts/IFilePathFinder.cs
@@ -1,0 +1,8 @@
+namespace ScriptCs.Contracts
+{
+    public interface IFilePathFinder
+    {
+        string[] FindPossibleAssemblyNames(string nameFragment, IFileSystem fileSystem);
+        string[] FindPossibleFilePaths(string pathFragment, IFileSystem fileSystem);
+    }
+}
